Added pooling of URL's and MySQL
URL's are now pooled for faster execution.
Data is now directly stored to a MySQL database.
As a result two new modules (multiprocessing and mysql.connector) are required
Web_crawler.py requires some parameters to connect to MySQL server.
(By default given as User = root,host = localhost and no password)
The Database name given must already exist in the MySQL server. 
Since the program checks if a given entry is already in the database. Crawling of webpage can be done in steps ( i.e If we try to crawl with steps = 50 first then steps = 100, the program will not fetch data for the first fifty ID's again as they were already fetched.Thus, in effect we crawled the first 50 ID's first then the remaining fifty ID's.)
verify_ssl set to False for fakeUserAgent as sometimes the certificate can't be authenticated leading to an error.